# Conflicts:
#	tests/common/HttpServer.hpp
#	tests/common/SocketTools.hpp
#	tests/unittests/HttpClientTests.cpp
#	tests/unittests/LoggerTests.cpp
#	tests/unittests/OfflineStorageTests_SQLiteWithMock.cpp
#	uap-cpp/uap-cpp.vcxproj

+++ tests/unittests/HttpServerTests.cpp
@@ -52,7 +52,7 @@ class HttpServerTestsBase : public ::testing::Test
 
     bool sendRequest(std::string const& request)
     {
-        return (clientSocket.send(request.data(), request.size()) == static_cast<int>(request.size()));
+        return (clientSocket.send(request.data(), static_cast<unsigned int>(request.size())) == static_cast<int>(request.size()));
     }
 
     std::string receiveResponse(bool singlePacket = false)
